Add some refactoring changes
